Improves expedition start flow  (#11)
The app now handles better waiting for the user to approve giving as location access via OS dialogues. 

We are now forcing the user to give us `AUTHORIZATION_STATUS_ALWAYS` but in the future we might have to do with whatever permission they give us BUT inform them that some features might not work.

This is good enough for now.